backend/requirements.txt
aiomysql==0.1.1
alembic==1.11.1
anyio==3.6.2
# async-generator==1.10
asyncmy
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
joblib==1.3.1
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
# mysql-connector-python==8.0.33
# mysqlclient==2.1.1
spacy
nltk==3.8.1
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.7
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
PyMySQL==1.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.8.8
requests==2.29.0
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.9.0
six==1.16.0
scikit-learn
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.14
starlette==0.27.0
threadpoolctl==3.2.0
tqdm==4.66.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
wsproto==1.2.0
XlsxWriter==3.1.0
